+++ pytests/test_testing.py
@@ -11,40 +11,40 @@ def test_ffwd_iter():
         next(it)
 
 
-def test_input():
+def test_input(now):
     inp = TestingSource(range(3))
-    part = inp.build_part("iterable", None)
-    assert part.next_batch() == [0]
-    assert part.next_batch() == [1]
-    assert part.next_batch() == [2]
+    part = inp.build_part(now, "iterable", None)
+    assert part.next_batch(now) == [0]
+    assert part.next_batch(now) == [1]
+    assert part.next_batch(now) == [2]
     with raises(StopIteration):
-        part.next_batch()
+        part.next_batch(now)
     part.close()
 
 
-def test_input_resume_state():
+def test_input_resume_state(now):
     inp = TestingSource(range(3))
-    part = inp.build_part("iterable", None)
-    assert part.next_batch() == [0]
-    assert part.next_batch() == [1]
+    part = inp.build_part(now, "iterable", None)
+    assert part.next_batch(now) == [0]
+    assert part.next_batch(now) == [1]
     resume_state = part.snapshot()
     assert resume_state == 2
-    assert part.next_batch() == [2]
+    assert part.next_batch(now) == [2]
     part.close()
 
     inp = TestingSource(range(3))
-    part = inp.build_part("iterable", resume_state)
+    part = inp.build_part(now, "iterable", resume_state)
     assert part.snapshot() == resume_state
-    assert part.next_batch() == [2]
+    assert part.next_batch(now) == [2]
     with raises(StopIteration):
-        part.next_batch()
+        part.next_batch(now)
     part.close()
 
 
-def test_input_batch_size():
+def test_input_batch_size(now):
     inp = TestingSource(range(5), batch_size=2)
-    part = inp.build_part("iterable", None)
-    assert part.next_batch() == [0, 1]
-    assert part.next_batch() == [2, 3]
-    assert part.next_batch() == [4]
+    part = inp.build_part(now, "iterable", None)
+    assert part.next_batch(now) == [0, 1]
+    assert part.next_batch(now) == [2, 3]
+    assert part.next_batch(now) == [4]
     part.close()